Move getExperimentAndBranchInfos out of pages
